route pay extra into the rest of the modal
