--- packages/web/src/components/premium-content-purchase-modal/components/FormatPrice.module.css
@@ -1,9 +1,0 @@
-.container {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--unit-2);
-}
-
-.existingBalance {
-  text-decoration: none;
-}
