+++ packages/web/src/components/premium-content-purchase-modal/components/PurchaseSummaryTable.tsx
@@ -2,32 +2,30 @@ import { formatPrice } from '@audius/common'
 
 import { Text } from 'components/typography'
 
-import { FormatPrice } from './FormatPrice'
 import styles from './PurchaseSummaryTable.module.css'
 
 const messages = {
   summary: 'Summary',
-  artistCut: 'Artist Cut',
-  audiusCut: 'Audius Cut',
-  alwaysZero: 'Always $0',
-  existingBalance: 'Existing Balance',
-  youPay: 'You Pay',
+  premiumTrack: 'Premium Track',
+  existingBalance: 'Existing USDC Balance',
+  payExtra: 'Pay Extra',
+  total: 'Total',
   youPaid: 'You Paid',
-  zero: '$0',
+  zero: '$0.00',
   price: (val: string) => `$${val}`
 }
 
 type PurchaseSummaryTableProps = {
   amountDue: number
-  artistCut: number
   basePrice: number
+  extraAmount?: number
   existingBalance?: number
   isPurchased?: boolean
 }
 
 export const PurchaseSummaryTable = ({
   amountDue,
-  artistCut,
+  extraAmount,
   basePrice,
   existingBalance,
   isPurchased
@@ -38,13 +36,15 @@ export const PurchaseSummaryTable = ({
         {messages.summary}
       </Text>
       <div className={styles.row}>
-        <span>{messages.artistCut}</span>
-        <span>{messages.price(formatPrice(artistCut))}</span>
+        <span>{messages.premiumTrack}</span>
+        <span>{messages.price(formatPrice(basePrice))}</span>
       </div>
-      <div className={styles.row}>
-        <span>{messages.audiusCut}</span>
-        <span>{messages.alwaysZero}</span>
-      </div>
+      {extraAmount != null ? (
+        <div className={styles.row}>
+          <span>{messages.payExtra}</span>
+          <span>{messages.price(formatPrice(extraAmount))}</span>
+        </div>
+      ) : null}
       {existingBalance ? (
         <div className={styles.row}>
           <span>{messages.existingBalance}</span>
@@ -52,12 +52,10 @@ export const PurchaseSummaryTable = ({
         </div>
       ) : null}
       <Text className={styles.row} variant='title'>
-        <span>{isPurchased ? messages.youPaid : messages.youPay}</span>
-        <FormatPrice
-          className={styles.finalPrice}
-          basePrice={basePrice}
-          amountDue={amountDue}
-        />
+        <span>{isPurchased ? messages.youPaid : messages.total}</span>
+        <span className={styles.finalPrice}>
+          {messages.price(formatPrice(amountDue))}
+        </span>
       </Text>
     </Text>
   )
